Rewrite to use utils from bootstrap & not require static offsets
* Switched to using utils from bootstrap.  dwarfs & unionfs
compilation, if needed is now performed as part of bootstrap instead
of a separate create-utils.sh which was removed.  Might create
PKGBUILD to serve same purpose as create-utils.sh in future to build
all tools statically if necessary.

* Instead of static offsets init.c was rewritten to parse header that
is concatenated at the beginning of script and performs extraction of
all conty files on it's own & conty-start.sh was modified to not do
any extraction.

* Remove requirement on being able to exec scripts from tmpfs by
moving all files to .local/share/conty

* Removed ability to set custom conty base directory name.  Might
bring it back in future but for now I just want to proceed with
modifying conty-start.sh without adding this functionality to init.c